Template for W/R DirectQuickSelectSketch split
